feat(SpatialTarget): allow gameobject to be passed to enter,exit,select
The SpatialTarget can now accept a GameObject as the parameter for
`Enter`, `Exit` and `Select` which will just make a basic SurfaceData
object to be passed into the normal methods. This allows the potential
for any object to trigger the spatial target states, such as an
Interactor to touch the spatial target and still affect the states.