feat(prefabs): allow object collisions to interact with target
The Spatial Target can now be interacted with by colliding objects
with it and the new optional hover collider allows for the hover
state to be activated when the colliding object is near the target.

The new collision states are only activated if the collidable objects
list is populated, otherwise it is ignored and the trigger collider
is disabled.