fix(prefabs): allow collision select after deselect occurs
There was an issue where the collision select would not be available
after the deselect was called because the select activation was only
happening on the LastExited event, meaning select can only happen
after everything stops hovering over it. But a deselect event should
also allow select to occur again.